+++ src/constants/solution-icon-images.ts
