+++ src/utils/merged-sector.ts
